test(personas): assert rendered avatar output instead of spying on PersonaAvatar
- actor-avatar.test.tsx now mounts the real components and asserts the
  rendered <img src> / emoji fallback (INV-39) instead of checking React's
  function-call signature (CodeRabbit review)
- extract the Radix-Avatar jsdom image-loading stub into a shared
  stubImageLoading() helper in @/test; persona-avatar.test.tsx reuses it
