+++ apps/frontend/src/components/actor-avatar.test.tsx
@@ -1,11 +1,16 @@
-import { render } from "@testing-library/react"
-import { describe, expect, it, vi } from "vitest"
+import { afterEach, beforeEach, describe, expect, it, vi } from "vitest"
+import { render, waitFor } from "@testing-library/react"
 import * as hooks from "@/hooks"
-import * as personaAvatarModule from "@/components/persona-avatar"
+import { stubImageLoading } from "@/test"
 import { ActorAvatar } from "./actor-avatar"
 
 describe("ActorAvatar persona branch", () => {
-  it("forwards the resolved avatarUrl to PersonaAvatar (custom persona images must reach the timeline)", () => {
+  beforeEach(() => {
+    stubImageLoading()
+  })
+  afterEach(() => vi.unstubAllGlobals())
+
+  it("renders the resolved avatar image (custom persona images must reach the timeline)", async () => {
     vi.spyOn(hooks, "useActors").mockReturnValue({
       getActorAvatar: () => ({
         fallback: "🐹",
@@ -13,17 +18,23 @@ describe("ActorAvatar persona branch", () => {
         avatarUrl: "/api/workspaces/ws_1/personas/persona_1/avatar/123.64.webp",
       }),
     } as unknown as ReturnType<typeof hooks.useActors>)
-    const personaAvatar = vi.spyOn(personaAvatarModule, "PersonaAvatar")
 
-    render(<ActorAvatar actorId="persona_1" actorType="persona" workspaceId="ws_1" />)
+    const { container } = render(<ActorAvatar actorId="persona_1" actorType="persona" workspaceId="ws_1" />)
 
-    expect(personaAvatar).toHaveBeenCalledWith(
-      expect.objectContaining({
-        slug: "stefan",
-        avatarUrl: "/api/workspaces/ws_1/personas/persona_1/avatar/123.64.webp",
-        fallback: "🐹",
-      }),
-      undefined
-    )
+    // The persona image is decorative (alt=""), so query the element directly.
+    await waitFor(() => {
+      const img = container.querySelector("img")
+      expect(img).toHaveAttribute("src", "/api/workspaces/ws_1/personas/persona_1/avatar/123.64.webp")
+    })
+  })
+
+  it("renders the emoji fallback when the persona has no image", () => {
+    vi.spyOn(hooks, "useActors").mockReturnValue({
+      getActorAvatar: () => ({ fallback: "🐹", slug: "stefan", avatarUrl: undefined }),
+    } as unknown as ReturnType<typeof hooks.useActors>)
+
+    const { getByText } = render(<ActorAvatar actorId="persona_1" actorType="persona" workspaceId="ws_1" />)
+
+    expect(getByText("🐹")).toBeInTheDocument()
   })
 })
